% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ stack }
\alias{ stack }
\alias{ stack }
\title{ Stack Data Types as an 'R6' Class }
\description{
An extremely simple stack data type, implemented with 'R6' classes. The size
  of the stack increases as needed, and the amortized time complexity is O(1).
  The stack may contain arbitrary objects.
}
